+++ src/titiler/xarray/titiler/xarray/extensions.py
@@ -1,19 +1,21 @@
 """titiler.xarray Extensions."""
 
 import warnings
-from typing import Callable, List, Type
+from typing import Annotated, Callable, Dict, List, Optional, Type
 
 import xarray
 from attrs import define
-from fastapi import Depends
+from fastapi import Depends, Query
+from rio_tiler.constants import WGS84_CRS
 from starlette.responses import HTMLResponse
+from typing_extensions import TypedDict
 
 from titiler.core.dependencies import DefaultDependency
 from titiler.core.factory import FactoryExtension
 from titiler.core.resources.enums import MediaType
 from titiler.xarray.dependencies import XarrayIOParams
 from titiler.xarray.factory import TilerFactory
-from titiler.xarray.io import open_zarr
+from titiler.xarray.io import X_DIM_NAMES, Y_DIM_NAMES, open_zarr
 
 
 @define
@@ -56,7 +58,7 @@ class DatasetMetadataExtension(FactoryExtension):
     io_dependency: Type[DefaultDependency] = XarrayIOParams
     dataset_opener: Callable[..., xarray.Dataset] = open_zarr
 
-    def register(self, factory: TilerFactory):
+    def register(self, factory: TilerFactory):  # noqa: C901
         """Register endpoint to the tiler factory."""
 
         @factory.router.get(
@@ -109,3 +111,118 @@ class DatasetMetadataExtension(FactoryExtension):
             """Returns the list of keys/variables in the Dataset."""
             with self.dataset_opener(src_path, **io_params.as_dict()) as ds:
                 return list(ds.data_vars)
+
+
+class ValidationInfo(TypedDict):
+    """Variable Validation model."""
+
+    compatible_with_titiler: bool
+    errors: List[str]
+    warnings: List[str]
+
+
+@define
+class ValidateExtension(FactoryExtension):
+    """Add /validate endpoints to a Xarray TilerFactory."""
+
+    io_dependency: Type[DefaultDependency] = XarrayIOParams
+    dataset_opener: Callable[..., xarray.Dataset] = open_zarr
+
+    def _validate_variable(self, da: xarray.DataArray) -> ValidationInfo:  # noqa: C901
+        errors: List[str] = []
+        warnings: List[str] = []
+
+        if len(da.dims) not in [2, 3]:
+            warnings.append(
+                f"DataArray has too many dimension ({len(da.dims)}) for titiler.xarray, dimensions reduction (sel) will be required.",
+            )
+
+        if "y" not in da.dims:
+            try:
+                y_dim = next(name for name in Y_DIM_NAMES if name in da.dims)
+                da = da.rename({y_dim: "y"})
+
+            except StopIteration:
+                errors.append(
+                    "Dataset does not have compatible `Y` spatial coordinates"
+                )
+
+        if "x" not in da.dims:
+            try:
+                x_dim = next(name for name in X_DIM_NAMES if name in da.dims)
+                da = da.rename({x_dim: "x"})
+            except StopIteration:
+                errors.append(
+                    "Dataset does not have compatible `X` spatial coordinates"
+                )
+
+        if {"x", "y"}.issubset(set(da.dims)):
+            if extra_dims := [d for d in da.dims if d not in ["x", "y"]]:
+                da = da.transpose(*extra_dims, "y", "x")
+            else:
+                da = da.transpose("y", "x")
+
+            bounds = da.rio.bounds()
+            if not bounds:
+                errors.append("Dataset does not have rioxarray bounds")
+
+            res = da.rio.resolution()
+            if not res:
+                errors.append("Dataset does not have rioxarray resolution")
+
+            if res and bounds:
+                crs = da.rio.crs or "epsg:4326"
+                xres, yres = map(abs, res)
+
+                # Adjust the longitude coordinates to the -180 to 180 range
+                if crs == "epsg:4326" and (da.x > 180 + xres / 2).any():
+                    da = da.assign_coords(x=(da.x + 180) % 360 - 180)
+
+                    # Sort the dataset by the updated longitude coordinates
+                    da = da.sortby(da.x)
+
+                bounds = tuple(da.rio.bounds())
+                if crs == WGS84_CRS and (
+                    bounds[0] + xres / 2 < -180
+                    or bounds[1] + yres / 2 < -90
+                    or bounds[2] - xres / 2 > 180
+                    or bounds[3] - yres / 2 > 90
+                ):
+                    errors.append(
+                        "Dataset bounds are not valid, must be in [-180, 180] and [-90, 90]"
+                    )
+
+            if not da.rio.transform():
+                errors.append("Dataset does not have rioxarray transform")
+
+        return {
+            "compatible_with_titiler": True if not errors else False,
+            "errors": errors,
+            "warnings": warnings,
+        }
+
+    def register(self, factory: TilerFactory):  # noqa: C901
+        """Register endpoint to the tiler factory."""
+
+        @factory.router.get(
+            "/validate",
+            responses={
+                200: {
+                    "content": {
+                        "application/json": {},
+                    },
+                },
+            },
+            response_model=Dict[str, ValidationInfo],
+        )
+        def validate_dataset(
+            src_path=Depends(factory.path_dependency),
+            io_params=Depends(self.io_dependency),
+            variables: Annotated[
+                Optional[List[str]], Query(description="Xarray Variable name.")
+            ] = None,
+        ):
+            """Returns the HTML representation of the Xarray Dataset."""
+            with self.dataset_opener(src_path, **io_params.as_dict()) as dst:
+                variables = variables or list(dst.data_vars)
+                return {v: self._validate_variable(dst[v]) for v in variables}